+++ client/src/App.tsx
@@ -1,8 +1,7 @@
-import { useState, useEffect } from "react";
 import "./App.scss";
 import Login from "./pages/login/Login";
 import Register from "./pages/register/Register";
-import { Route, BrowserRouter, Routes, redirect } from "react-router-dom";
+import { Route, BrowserRouter, Routes } from "react-router-dom";
 import { AuthContextProvider } from "./context/Auth";
 import { ChakraProvider } from "@chakra-ui/react";
 import Home from "./pages/home/Home";
@@ -12,25 +11,8 @@ import EditarAluno from "./pages/editarAluno/EditarAluno";
 import Metricas from "./pages/metricas/Metricas";
 import { AuthRoute } from "./components/authRoute/AuthRoute";
 import NotFound from "./pages/notFound/NotFound";
-import { AsyncLocalStorage } from "./util/AsyncLocalStorage";
 
 function App() {
-  const [userFromStorage, setUserFromStorage] = useState<object | null>(null);
-
-  useEffect(() => {
-    const getUserFromStorage = async () => {
-      try {
-        const getUser = await AsyncLocalStorage.getItem("userData");
-        const parseUser = JSON.parse(getUser!)
-        setUserFromStorage(parseUser);
-        console.log(parseUser)
-         userFromStorage ? redirect("/home") : redirect("/")
-      } catch (error) {
-        console.error(error);
-      }
-    }
-    getUserFromStorage()
-  }, []);
 
   return (
     <ChakraProvider>
